Move GetTempPath() to testutil.
